+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [v1.6.3]
 - added session-id to request response message [#454](https://github.com/xmidt-org/webpa-common/pull/454)
+- added wrp-go v2.0.0 to device package [#454](https://github.com/xmidt-org/webpa-common/pull/454)
 
 ## [v1.6.2]
 - Updated capabilityChecker to include metrics and configure whether or not to return errors [#449](https://github.com/xmidt-org/webpa-common/pull/449)